pyproject.toml
[project]
name = "n8n"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord.py==2.6.3",
    "httpx==0.28.1"
]

[project.optional-dependencies]
dev = []